tests/wpt/tests/css/css-ui/parsing/field-sizing-valid.html

<!DOCTYPE html>
<title>CSS UI: parsing field-sizing with valid values</title>
<link rel="help" href="https://drafts.csswg.org/css-ui-4/#field-sizing">
<link rel="author" title="Kent Tamura" href="mailto:tkent@chromium.org">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
<body>
<script>
test_valid_value('field-sizing', 'fixed');
test_valid_value('field-sizing', 'content');
</script>
</body>
